Move check test into `C1_verification`
