+++ tests/test_C0_data_processing.py
@@ -57,7 +57,6 @@ from multi_vector_simulator.utils.constants_json_strings import (
 )
 from multi_vector_simulator.utils.exceptions import (
     InvalidPeakDemandPricingPeriodsError,
-    UnknownEnergyCarrierError,
 )
 
 
@@ -443,24 +442,6 @@ def test_evaluate_lifetime_costs():
         SPECIFIC_REPLACEMENT_COSTS_OPTIMIZED,
     ]:
         assert k in dict_asset, f"Function does not add {k} to the asset dictionary."
-
-
-def test_check_if_energy_carrier_is_defined_in_DEFAULT_WEIGHTS_ENERGY_CARRIERS_pass():
-    # Function only needs to pass
-    C0.check_if_energy_carrier_is_defined_in_DEFAULT_WEIGHTS_ENERGY_CARRIERS(
-        "Electricity", "asset_group", "asset"
-    )
-    assert (
-        1 == 1
-    ), f"The energy carrier `Electricity` is not recognized to be defined in `DEFAULT_WEIGHTS_ENERGY_CARRIERS`."
-
-
-def test_check_if_energy_carrier_is_defined_in_DEFAULT_WEIGHTS_ENERGY_CARRIERS_fail():
-    with pytest.raises(UnknownEnergyCarrierError):
-        C0.check_if_energy_carrier_is_defined_in_DEFAULT_WEIGHTS_ENERGY_CARRIERS(
-            "Bio-Diesel", "asset_group", "asset"
-        ), f"The energy carrier `Bio-Diesel` is recognized in the `DEFAULT_WEIGHTS_ENERGY_CARRIERS`, eventhough it should not be defined."
-
 
 group = "GROUP"
 asset = "ASSET"
